522: v1.1: add lti and do better logger
